Refactor: Rename "Current Post Details" to "Performance Details"
